Remake the messages/get api to messages/firstvisit and add chart date fetching part
